+++ changelog.md
@@ -1,7 +1,16 @@
+### 1.7:
+  - Updated plugin description
+  - Updated README
+  - Updated Config
+  - Moved BlacklistCommand to Main
+  - Removed PlayerColorCodes and PlayerLogger
+  - Updated Logger
+  - Optimized Strings formatting 
+
 #### 1.6:
   - **CRITICAL**: Removed IP and subnet support, because it's not the goal of this plugin
   - Changed logger system
-  - Removed VersionChecker because it's useless for type of plugin
+  - Removed VersionChecker because it's useless for this type of plugin
   - Changed command usage
   - Updated README and added a warning to use the [Anti-VPN-Service](github.com/xpdustry/Anti-VPN-Service) plugin for IP and subnet filtering
   - Added colors when displaying plugin config